Update index.html
add a checkbox at the top of each Adquirente and Transmitente section. When checked, it will copy the user's data into the form fields. The data can still be modified after copying. Additionally, if the user changes their own data after copying it to an Adquirente or Transmitente section, the copied data will be updated automatically if the checkbox is still checked.